Remove deprecated tempdir dependency
`tempdir` had a dependency on `remove_dir_all` with was subject to a CWE

https://github.com/advisories/GHSA-mc8h-8q98-g5hr

As `tempdir` is deprecated this commit switches to using `tempfile`
which contains much of the functionality of `tempfile`. As this will
change the location in which some of the files are created this commit
also bumps the minor version of the crate.