feat(internal): Replace `cp` use with system call
Previously the flags we provided to cp were compiled into the library
based on platform, this caused issues when using a different binary to
the one provided by the system as flags were sometimes not available.

This commit adds a dependency on reflink-copy to enable the same copy on
write behavior for postgresql directories but without using the cp
binary directly.