feat: Allow migrations in the cache dir
When using this crate, running migrations multiple times can be quite
painful. This commit adds the option to use a callback to run migrations
when initialising the factory. Combined with copy on write for the cache
directory this significantly speeds up test suites. Right now this is
only available for synchronous functions but an asynchronous version
could be added later.